pyproject.toml
[project]
name = "butler-agent"
version = "0.1.0"
description = "AI-powered conversational DevOps agent for Kubernetes infrastructure management"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [
    {name = "OSDU Community", email = "noreply@osdu.org"}
]
keywords = ["kubernetes", "devops", "ai", "agent", "kind", "gitops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # Core dependencies
    "agent-framework>=1.0.0b251007",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "prompt_toolkit>=3.0.0",
    "typer>=0.12.0",
    "kubernetes>=31.0.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    # LLM Provider SDKs
    "openai>=1.58.1",
    "azure-identity>=1.25.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "mypy>=1.18.2",
    "black>=25.9.0",
    "ruff>=0.13.3",
]

[project.scripts]
butler = "agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.isort]
known-first-party = ["agent"]

[tool.coverage.run]
source = ["src/agent"]
omit = [
    "tests/*",
    "src/agent/display/*",
    "src/agent/observability.py",
    "src/agent/cli.py",
    "src/agent/cluster/kind_manager.py",
    "src/agent/cluster/status.py",
    "src/agent/cluster/tools.py",
    "src/agent/utils/paths.py",
    "src/agent/utils/validation.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
